refactor: move eval.py to src/miner directory
